requirements.txt
flask
gunicorn
marshmallow
sqlalchemy
flask_jwt_extendedclr